Changed: Remove dafult dropdown behavior helper methods from API.
